Create a Config class
The MicroMVC function `config()` is in a few classes in Micro. We should replace it with a config class as part of Micro which MicroMVC can then use. Any classes that need the config functionality can make sure to get it using dependency injection (constructor or setter, depending on the implementation details).